Fix "Not declaring uint as uint256"

+++ eth-contracts/contracts/ClaimsManager.sol
@@ -25,7 +25,7 @@ contract ClaimsManager is InitializableV2 {
       *       Avg block time - 13s
       *       604800 / 13 = 46523.0769231 blocks
       */
-    uint private fundingRoundBlockDiff;
+    uint256 private fundingRoundBlockDiff;
 
     /**
       * @notice - Configures the current funding amount per round
@@ -36,10 +36,10 @@ contract ClaimsManager is InitializableV2 {
       * @dev - Past a certain block height, this schedule will be updated
       *      - Logic determining schedule will be sourced from an external contract
       */
-    uint private fundingAmount;
+    uint256 private fundingAmount;
 
     // Denotes current round
-    uint private roundNumber;
+    uint256 private roundNumber;
 
     // Staking contract ref
     ERC20Mintable private audiusToken;
@@ -49,25 +49,25 @@ contract ClaimsManager is InitializableV2 {
     // 2) Total funded for this round
     // 3) Total claimed in round
     struct Round {
-        uint fundBlock;
-        uint fundingAmount;
-        uint totalClaimedInRound;
+        uint256 fundBlock;
+        uint256 fundingAmount;
+        uint256 totalClaimedInRound;
     }
 
     // Current round information
     Round currentRound;
 
     event RoundInitiated(
-      uint _blockNumber,
-      uint _roundNumber,
-      uint _fundAmount
+      uint256 _blockNumber,
+      uint256 _roundNumber,
+      uint256 _fundAmount
     );
 
     event ClaimProcessed(
       address _claimer,
-      uint _rewards,
-      uint _oldTotal,
-      uint _newTotal
+      uint256 _rewards,
+      uint256 _oldTotal,
+      uint256 _newTotal
     );
 
     /**
@@ -99,25 +99,25 @@ contract ClaimsManager is InitializableV2 {
     }
 
     /// @notice Get the duration of a funding round in blocks
-    function getFundingRoundBlockDiff() external view returns (uint blockDiff)
+    function getFundingRoundBlockDiff() external view returns (uint256 blockDiff)
     {
         return fundingRoundBlockDiff;
     }
 
     /// @notice Get the last block where a funding round was initiated
-    function getLastFundBlock() external view returns (uint lastFundBlock)
+    function getLastFundBlock() external view returns (uint256 lastFundBlock)
     {
         return currentRound.fundBlock;
     }
 
     /// @notice Get the amount funded per round in wei
-    function getFundsPerRound() external view returns (uint amount)
+    function getFundsPerRound() external view returns (uint256 amount)
     {
         return fundingAmount;
     }
 
     /// @notice Get the total amount claimed in the current round
-    function getTotalClaimedInRound() external view returns (uint claimedAmount)
+    function getTotalClaimedInRound() external view returns (uint256 claimedAmount)
     {
         return currentRound.totalClaimedInRound;
     }
@@ -226,7 +226,7 @@ contract ClaimsManager is InitializableV2 {
      */
     function processClaim(
         address _claimer,
-        uint _totalLockedForSP
+        uint256 _totalLockedForSP
     ) external
     {
         _requireIsInitialized();
@@ -237,9 +237,9 @@ contract ClaimsManager is InitializableV2 {
 
         Staking stakingContract = Staking(stakingAddress);
         // Prevent duplicate claim
-        uint lastUserClaimBlock = stakingContract.lastClaimedFor(_claimer);
+        uint256 lastUserClaimBlock = stakingContract.lastClaimedFor(_claimer);
         require(lastUserClaimBlock <= currentRound.fundBlock, "Claim already processed for user");
-        uint totalStakedAtFundBlockForClaimer = stakingContract.totalStakedForAt(
+        uint256 totalStakedAtFundBlockForClaimer = stakingContract.totalStakedForAt(
             _claimer,
             currentRound.fundBlock);
 
@@ -249,11 +249,11 @@ contract ClaimsManager is InitializableV2 {
 
         // Once they claim the zero reward amount, stake can be modified once again
         // Subtract total locked amount for SP from stake at fund block
-        uint claimerTotalStake = totalStakedAtFundBlockForClaimer.sub(_totalLockedForSP);
-        uint totalStakedAtFundBlock = stakingContract.totalStakedAt(currentRound.fundBlock);
+        uint256 claimerTotalStake = totalStakedAtFundBlockForClaimer.sub(_totalLockedForSP);
+        uint256 totalStakedAtFundBlock = stakingContract.totalStakedAt(currentRound.fundBlock);
 
         // Calculate claimer rewards
-        uint rewardsForClaimer = (
+        uint256 rewardsForClaimer = (
           claimerTotalStake.mul(fundingAmount)
         ).div(totalStakedAtFundBlock);
 
@@ -278,7 +278,7 @@ contract ClaimsManager is InitializableV2 {
         currentRound.totalClaimedInRound = currentRound.totalClaimedInRound.add(rewardsForClaimer);
 
         // Update round claim value
-        uint newTotal = stakingContract.totalStakedFor(_claimer);
+        uint256 newTotal = stakingContract.totalStakedFor(_claimer);
 
         emit ClaimProcessed(
             _claimer,
@@ -292,8 +292,8 @@ contract ClaimsManager is InitializableV2 {
      * @notice Modify funding amount per round
      * @param _newAmount - new amount to fund per round in wei
      */
-    function updateFundingAmount(uint _newAmount)
-    external returns (uint newAmount)
+    function updateFundingAmount(uint256 _newAmount)
+    external returns (uint256 newAmount)
     {
         require(
             msg.sender == governanceAddress,
@@ -310,8 +310,8 @@ contract ClaimsManager is InitializableV2 {
      * @return boolean - true if eligible for claim, false if not
      */
     function claimPending(address _sp) external view returns (bool pending) {
-        uint lastClaimedForSP = Staking(stakingAddress).lastClaimedFor(_sp);
-        (,,,uint numEndpoints,,) = (
+        uint256 lastClaimedForSP = Staking(stakingAddress).lastClaimedFor(_sp);
+        (,,,uint256 numEndpoints,,) = (
             ServiceProviderFactory(serviceProviderFactoryAddress).getServiceProviderDetails(_sp)
         );
         return (lastClaimedForSP < currentRound.fundBlock && numEndpoints > 0);
@@ -321,7 +321,7 @@ contract ClaimsManager is InitializableV2 {
      * @notice Modify minimum block difference between funding rounds
      * @param _newFundingRoundBlockDiff - new min block difference to set
      */
-    function updateFundingRoundBlockDiff(uint _newFundingRoundBlockDiff) external {
+    function updateFundingRoundBlockDiff(uint256 _newFundingRoundBlockDiff) external {
         require(
             msg.sender == governanceAddress,
             "Only callable by Governance contract"
