+++ eth-contracts/contracts/ServiceTypeManager.sol
@@ -23,8 +23,8 @@ contract ServiceTypeManager is InitializableV2 {
 
     /// @dev Struct representing service type stake requirements
     struct ServiceTypeStakeRequirements {
-        uint minStake;
-        uint maxStake;
+        uint256 minStake;
+        uint256 maxStake;
     }
 
     /// @dev mapping of service type to registered requirements
@@ -70,8 +70,8 @@ contract ServiceTypeManager is InitializableV2 {
      */
     function addServiceType(
         bytes32 _serviceType,
-        uint _serviceTypeMin,
-        uint _serviceTypeMax
+        uint256 _serviceTypeMin,
+        uint256 _serviceTypeMax
     ) external
     {
         _requireIsInitialized();
@@ -95,9 +95,9 @@ contract ServiceTypeManager is InitializableV2 {
 
         require(msg.sender == governanceAddress, "Only callable by Governance contract");
 
-        uint serviceIndex = 0;
+        uint256 serviceIndex = 0;
         bool foundService = false;
-        for (uint i = 0; i < validServiceTypes.length; i ++) {
+        for (uint256 i = 0; i < validServiceTypes.length; i ++) {
             if (validServiceTypes[i] == _serviceType) {
                 serviceIndex = i;
                 foundService = true;
@@ -106,7 +106,7 @@ contract ServiceTypeManager is InitializableV2 {
         }
         require(foundService == true, "Invalid service type, not found");
         // Overwrite service index
-        uint lastIndex = validServiceTypes.length - 1;
+        uint256 lastIndex = validServiceTypes.length - 1;
         validServiceTypes[serviceIndex] = validServiceTypes[lastIndex];
         validServiceTypes.length--;
         // Overwrite values
@@ -122,8 +122,8 @@ contract ServiceTypeManager is InitializableV2 {
      */
     function updateServiceType(
         bytes32 _serviceType,
-        uint _serviceTypeMin,
-        uint _serviceTypeMax
+        uint256 _serviceTypeMin,
+        uint256 _serviceTypeMax
     ) external
     {
         _requireIsInitialized();
@@ -144,7 +144,7 @@ contract ServiceTypeManager is InitializableV2 {
      * @return min and max stake for type
      */
     function getServiceTypeStakeInfo(bytes32 _serviceType)
-    external view returns (uint min, uint max)
+    external view returns (uint256 min, uint256 max)
     {
         return (
             serviceTypeStakeRequirements[_serviceType].minStake,
@@ -205,7 +205,7 @@ contract ServiceTypeManager is InitializableV2 {
      * @param _serviceType - type of service
      * @param _versionIndex - index in list of service versions
      */
-    function getVersion(bytes32 _serviceType, uint _versionIndex)
+    function getVersion(bytes32 _serviceType, uint256 _versionIndex)
     external view returns (bytes32 version)
     {
         require(
@@ -227,7 +227,7 @@ contract ServiceTypeManager is InitializableV2 {
             serviceTypeVersions[_serviceType].length >= 1,
             "No registered version of serviceType"
         );
-        uint latestVersionIndex = serviceTypeVersions[_serviceType].length - 1;
+        uint256 latestVersionIndex = serviceTypeVersions[_serviceType].length - 1;
         return (serviceTypeVersions[_serviceType][latestVersionIndex]);
     }
 
@@ -236,7 +236,7 @@ contract ServiceTypeManager is InitializableV2 {
      * @param _serviceType - type of service
      */
     function getNumberOfVersions(bytes32 _serviceType)
-    external view returns (uint)
+    external view returns (uint256)
     {
         return serviceTypeVersions[_serviceType].length;
     }
