+++ eth-contracts/contracts/test/MockStakingCaller.sol
@@ -10,8 +10,8 @@ import "../AudiusAdminUpgradeabilityProxy.sol";
 // Forwards basic staking functions
 // Forwards ServiceProviderFactory functions as well
 contract MockStakingCaller is InitializableV2 {
-    uint max;
-    uint min;
+    uint256 max;
+    uint256 min;
     using SafeERC20 for ERC20;
     Staking staking = Staking(0);
     ERC20 internal stakingToken;
@@ -35,7 +35,7 @@ contract MockStakingCaller is InitializableV2 {
 
     // Test only function
     function stakeRewards(
-        uint _amount,
+        uint256 _amount,
         address _staker
     ) external {
         _requireIsInitialized();
@@ -80,12 +80,12 @@ contract MockStakingCaller is InitializableV2 {
     /// @notice Calculate the stake for an account based on total number of registered services
     function getServiceProviderDetails(address)
     external view returns (
-        uint deployerStake,
-        uint deployerCut,
+        uint256 deployerStake,
+        uint256 deployerCut,
         bool validBounds,
-        uint numberOfEndpoints,
-        uint minAccountStake,
-        uint maxAccountStake)
+        uint256 numberOfEndpoints,
+        uint256 minAccountStake,
+        uint256 maxAccountStake)
     {
         return (0, 0, withinBounds, 1, min, max);
     }
